backend/scripts/pyproject.toml
[tool.poetry]
name = "scripts"
version = "0.1.0"
description = ""
authors = ["serramatutu <lucasvvop@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
websocket = "^0.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
